+++ src/decomp/c_0803CA9C.c
@@ -17,7 +17,7 @@ u8 sub_0803CA9C(u32 id)
     return (1 << (id & 7)) & *p;
 }
 
-u8 sub_0803CAB8(u32 id)
+int sub_0803CAB8(u32 id)
 {
     struct Unk02028030 *s = &gUnknown_02028030;
     u32 idx = id >> 3;
